AI-project/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
firebase-admin==6.4.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
gunicorn==21.2.0